Add unit to recommended specs and hide custom deployment button
The changes made are primarily UI improvements in the problem_id page. Units (core for CPU, GB for RAM and Disk) were added to the recommended specifications to clarify what the figures represent to the users. Additionally, the button for custom specification deployment was commented out. The decision to hide this button was driven by the need to simplify the decision making process for users in their early interactions with the app. It can be reintroduced once users are more familiar with the system and are more likely to benefit from advanced options.

Signed-off-by: Hsuan <34859514+hsuan1117@users.noreply.github.com>